# Remove HubProtocol minor version