--- a/src/SignalR/common/SignalR.Common/src/Protocol/HandshakeProtocol.cs
+++ b/src/SignalR/common/SignalR.Common/src/Protocol/HandshakeProtocol.cs
@@ -23,8 +23,6 @@ namespace Microsoft.AspNetCore.SignalR.Protocol
         private static ReadOnlySpan<byte> ProtocolPropertyNameBytes => new byte[] { (byte)'p', (byte)'r', (byte)'o', (byte)'t', (byte)'o', (byte)'c', (byte)'o', (byte)'l' };
         private const string ProtocolVersionPropertyName = "version";
         private static ReadOnlySpan<byte> ProtocolVersionPropertyNameBytes => new byte[] { (byte)'v', (byte)'e', (byte)'r', (byte)'s', (byte)'i', (byte)'o', (byte)'n' };
-        private const string MinorVersionPropertyName = "minorVersion";
-        private static ReadOnlySpan<byte> MinorVersionPropertyNameBytes => new byte[] { (byte)'m', (byte)'i', (byte)'n', (byte)'o', (byte)'r', (byte)'V', (byte)'e', (byte)'r', (byte)'s', (byte)'i', (byte)'o', (byte)'n' };
         private const string ErrorPropertyName = "error";
         private static ReadOnlySpan<byte> ErrorPropertyNameBytes => new byte[] { (byte)'e', (byte)'r', (byte)'r', (byte)'o', (byte)'r' };
         private const string TypePropertyName = "type";
@@ -40,7 +38,7 @@ namespace Microsoft.AspNetCore.SignalR.Protocol
                 var memoryBufferWriter = MemoryBufferWriter.Get();
                 try
                 {
-                    WriteResponseMessage(new HandshakeResponseMessage(protocol.MinorVersion), memoryBufferWriter);
+                    WriteResponseMessage(HandshakeResponseMessage.Empty, memoryBufferWriter);
                     result = memoryBufferWriter.ToArray();
                     _messageCache.TryAdd(protocol, result);
                 }
@@ -100,8 +98,6 @@ namespace Microsoft.AspNetCore.SignalR.Protocol
                     writer.WriteString(ErrorPropertyNameBytes, responseMessage.Error);
                 }
 
-                writer.WriteNumber(MinorVersionPropertyNameBytes, responseMessage.MinorVersion);
-
                 writer.WriteEndObject();
                 writer.Flush();
                 Debug.Assert(writer.CurrentDepth == 0);
@@ -133,7 +129,6 @@ namespace Microsoft.AspNetCore.SignalR.Protocol
             reader.CheckRead();
             reader.EnsureObjectStart();
 
-            int? minorVersion = null;
             string error = null;
 
             while (reader.CheckRead())
@@ -150,10 +145,6 @@ namespace Microsoft.AspNetCore.SignalR.Protocol
                     {
                         error = reader.ReadAsString(ErrorPropertyName);
                     }
-                    else if (reader.TextEquals(MinorVersionPropertyNameBytes))
-                    {
-                        minorVersion = reader.ReadAsInt32(MinorVersionPropertyName);
-                    }
                     else
                     {
                         reader.Skip();
@@ -169,7 +160,7 @@ namespace Microsoft.AspNetCore.SignalR.Protocol
                 }
             };
 
-            responseMessage = new HandshakeResponseMessage(minorVersion, error);
+            responseMessage = new HandshakeResponseMessage(error);
             return true;
         }