Fixes: #5320 
We introduced the HubProtocol minor version in 3.0. We don't plan on using it (we can add it back later if we change our minds) and it's a breaking change so let's remove it,