@mikaelm12 2019-05-02
Some benchmarking data for this change. Some minor improvements on Handshake.

Before

Method |     Mean |     Error |   StdDev |      Op/s |  Gen 0 | Allocated |
---------------------- |---------:|----------:|---------:|----------:|-------:|----------:|
 SuccessHandshakeAsync | 6.385 us | 0.6681 us | 1.970 us | 156,606.9 | 0.1297 |     680 B |

After

Method |     Mean |     Error |   StdDev |      Op/s |  Gen 0 | Allocated |
---------------------- |---------:|----------:|---------:|----------:|-------:|----------:|
 SuccessHandshakeAsync | 6.340 us | 0.5179 us | 1.519 us | 157,723.7 | 0.1526 |     648 B |